lms/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
dj-static==0.0.6
Django==3.1.3
django-celery-beat==2.1.0
django-cors-headers==3.3.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
flower==0.9.5
gunicorn==19.7.1
humanize==3.1.0
idna==2.9
importlib-metadata==3.1.0
kombu==4.6.11
oauthlib==3.1.0
Pillow==8.0.1
prometheus-client==0.8.0
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2020.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
static3==0.7.0
tornado==6.1
tqdm==4.54.0
urllib3==1.25.9
vine==5.0.0
zipp==3.4.0

scikit-learn~=0.23.2